backend/requirements.txt
asgiref==3.8.1
dataclasses==0.6
dj-rest-auth==7.0.1
Django==5.1.7
django-allauth==65.4.1
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.5.0
djongo==1.2.31
dnspython==2.7.0
drf-nested-routers==0.94.1
pillow==11.1.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
pymongo==4.11.2
python-dotenv==1.0.1
pytz==2025.1
sqlparse==0.5.3
